+++ package.json
@@ -13,10 +13,11 @@
     "format": "prettier --write ."
   },
   "dependencies": {
-    "@chakra-ui/react": "2.8.0",
-    "@emotion/react": "11",
-    "@emotion/styled": "11",
-    "framer-motion": "6",
+    "@chakra-ui/icons": "^2.1.1",
+    "@chakra-ui/react": "^2.8.2",
+    "@emotion/react": "^11.11.1",
+    "@emotion/styled": "^11.11.0",
+    "framer-motion": "^10.16.16",
     "i18next": "22.4.15",
     "next": "12.1.6",
     "next-i18next": "13.2.2",
