+++ src/modules/LandingPage/LandingPageContainer.tsx
@@ -1,0 +1,11 @@
+import { Box } from '@chakra-ui/react';
+
+import Layout from '@/uikit/Layout';
+
+export function LandingPageContainer() {
+  return (
+    <Layout>
+      <Box />
+    </Layout>
+  );
+}
