+++ src/pages/home.tsx
@@ -1,0 +1,26 @@
+import type { GetStaticPropsContext } from 'next';
+import { i18n } from 'next-i18next';
+import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
+import getConfig from 'next/config';
+
+import { LandingPageContainer } from '@/modules/LandingPage/LandingPageContainer';
+
+const { publicRuntimeConfig } = getConfig();
+
+function LandingPage() {
+  return <LandingPageContainer />;
+}
+
+export async function getStaticProps({ locale }: GetStaticPropsContext) {
+  if (!publicRuntimeConfig.isProduction) {
+    await i18n?.reloadResources();
+  }
+
+  return {
+    props: {
+      ...(locale ? await serverSideTranslations(locale, ['common']) : {}),
+    },
+  };
+}
+
+export default LandingPage;
